Restructure configuration
* Extract preact filter and esbuild into plugins
* Move URLs for spreadsheets into global variables file
* Make preact plugin more reusable by adding configuration for component location
