+++ _config/globals.js
@@ -1,0 +1,4 @@
+export default {
+    truthSocialCsvUrl: "https://docs.google.com/spreadsheets/d/1vbsoLq-Z5_kJaV0GOFMOqyo3dL9S1SKfUSkNWdYtMtU/export?format=csv&id=1vbsoLq-Z5_kJaV0GOFMOqyo3dL9S1SKfUSkNWdYtMtU&gid=201966548",
+    twitterCsvUrl: "https://docs.google.com/spreadsheets/d/1vbsoLq-Z5_kJaV0GOFMOqyo3dL9S1SKfUSkNWdYtMtU/export?format=csv&id=1vbsoLq-Z5_kJaV0GOFMOqyo3dL9S1SKfUSkNWdYtMtU&gid=0",
+}
