+++ src/_components/post-explorer.webc
@@ -1,5 +1,5 @@
 <div @raw="preact(
-    './src/preact/components/post-explorer.js',
+    'post-explorer.js',
     { posts: posts, postsPerPage: 25, initialRange: [3, 25] }
 )" webc:nokeep></div>
