+++ src/_components/post-full.webc
@@ -1,5 +1,5 @@
 <div @raw="preact(
-    './src/preact/components/post-full.js',
+    'post-full.js',
     { post: post, horizontal: horizontal }
 )"></div>
