+++ src/_data/posts.js
@@ -2,9 +2,6 @@ import Fetch from '@11ty/eleventy-fetch'
 import Papa from 'papaparse'
 import crypto from 'crypto'
 
-const TRUTH_SOCIAL_URL = 'https://docs.google.com/spreadsheets/d/1vbsoLq-Z5_kJaV0GOFMOqyo3dL9S1SKfUSkNWdYtMtU/export?format=csv&id=1vbsoLq-Z5_kJaV0GOFMOqyo3dL9S1SKfUSkNWdYtMtU&gid=201966548'
-const TWITTER_URL = 'https://docs.google.com/spreadsheets/d/1vbsoLq-Z5_kJaV0GOFMOqyo3dL9S1SKfUSkNWdYtMtU/export?format=csv&id=1vbsoLq-Z5_kJaV0GOFMOqyo3dL9S1SKfUSkNWdYtMtU&gid=0'
-
 function hashString(str) {
     const hash = crypto.createHash('sha256')
     hash.update(str)
@@ -57,8 +54,8 @@ function splitPreservingQuotes(str) {
     return result.filter(item => item !== '')
 }
 
-export default async function () {
-    let truthCSV = await Fetch(TRUTH_SOCIAL_URL, {
+export default async function (configData) {
+    let truthCSV = await Fetch(configData.truthSocialCsvUrl, {
         duration: "1d",
         type: "text"
     })
@@ -89,7 +86,7 @@ export default async function () {
             }
         }).filter(d => d.content !== '')
 
-    let twitterCSV = await Fetch(TWITTER_URL, {
+    let twitterCSV = await Fetch(configData.twitterCsvUrl, {
         duration: "1d",
         type: "text"
     })
